RGB/VGA: Major Error with runcommand-onstart.sh
This fixes a MAJOR issue with the RGB and VGA forks where the script that looks for values in 320.txt or 240.txt was forcing everything into 2048x240p, regardless of the rest of the script.

This fixes that, and allows proper use of the two text files to force resolutions.